+++ code/python_tooling/dashboard/server.py
@@ -9,3 +9,7 @@ app = Dash(title='Reprojection', update_title=None)
 # TODO(Jack): Do not hardcode this - giving a user the ability to interact with the file system in a gui is not so
 #  trivial but can be done with some tk tools or other libraries
 DB_DIR = '../../test_data/'
+
+
+# TODO(Jack): Place meta data like this in config file or in database. For now we use globals...
+IMAGE_DIMENSIONS = (512, 512)
